Souce, Target, Dest Tanks and local fuel transfer
Some quirkyness with the page selection, ensure only ONE button is
pressed at a time other than the Online button to ensure you are on the
correct page.

The transfer window should only be used after you have selected a source
and destination tank. It currently transfer it instantly, but this
should eventually be based upon some flow rate variable.

Current MAX targeting distance is hard coded at 100,000 although it is
not currently working. This is just so that I don't have to spend
forever getting them close enough.