Massive overhaul to the List Vessels Screen and Code Refactor
Removed a lot of the wrapper data structures to simplify the code.

Reworked the display of the List Vessels screen.
The meanings of the colors are as follows:
White == distance beyond 100km
Yellow = distance < 100km but beyond 2km
Green == distance < 2km, simulation active. Can scan for fueltanks.
We now can transfer fuel if we want.